Revert "Update for rigify changes in 2.81 and 2.82"
